get rid of the signature stuff and generalize the handler method
We didn't really use the Signature stuff anyway and it's easier to
generalize such logic in each classes their handles() method. Also
introduces a magic field to which the handles() method defaults, next to
the picker() method. Furthermore, this commit simplifies the detection
logic in the tarfile and zip classes.