be less strict about empty/corrupt archives
In practice it's very possible to run into an empty or corrupted archive
container. In those cases we really don't want to be throwing around
exceptions that undo any other unpacking success that may have happened.
This commit therefore smooths the handling of such archive files.

Note that some of the try/catch statements in the unpackers have a
fallthrough mechanism. This mechanism makes sure that any temporarily
created directories are deleted.