properly handle a certain .msg file
This commit fixes two issues. First of all, apparently the .msg file
only has an ASCII filename specification for the embedded .doc file and
as such the .decode() for the Unicode filename is throwing an exception.
Secondly, the .doc filename has a trailing null-byte, potentially
causing issues with our unpacking mechanism.

So in our Cuckoo Sandbox use-case we transform the .msg into a .zip and
submit it to Cuckoo with the filename of our target as parameter; now
for the extraction part (extracting the .doc from the .zip) it's
necessary to include the trailing null-byte as otherwise we simply get a
404 file not found back from our OLE unpacker. However, when submitting
the analysis to Cuckoo, the filename should not have the trailing
null-byte for several obvious reasons:
* .zip files don't support trailing null-bytes (zipfile strips 'em)
* SQL databases don't like null-bytes (PostgreSQL seems to strip it)
* NTFS & Python don't like null-bytes either

This leaves us with the following solution/workaround: keep the trailing
null-bytes in the "extrpath" and "relapath" fields and strip 'em off
from the "filename" field. This seems to provide the desired outcome.