ensure duplicates is the original list
This isn't a direct bug currently (as it was *also* fixed by the
previous commit), but just making sure. In the case where a newly
initiated duplicates list is passed along to an archive, the list may
end up empty and therefore re-initialized. In effect any files that are
added down the road won't show up in the global duplicates list - as a
local shadow was made of it. Therefore preventing duplicate detection
from doing its job.