allow package definition for unpackers, some tweaks
PeePDF never returns an error code (-1) in the parse() method, so don't
check for that. If anything, we may want to take a look at the f.errors
list, but presumably that will be full of false positives (because PDF).

Also ensure that an Unpacker's unpack() never returns None. One should
assume Unpacker.unpack() returns a list.