ci on py35 py36, remove zip password encryption
Adds Continuous Integration for Python 3.5 & Python 3.6, after which it
also turned out that on Python 3.6 the Zip password encryption wasn't
working anymore. As we weren't using this feature anyway, this commit
removes it altogether.
Also squeezes together the various py.test invocations such that only
two are required in total (which in practice doesn't matter).