ship win64 binaries for libmagic
Thanks to Sean Gillespie [1]. Now sflock supports magic(5) on each
platform that Cuckoo supports we can finally use it directly (as opposed
to the duplicate cuckoo.compat.magic layer that Cuckoo features).

The other reason Cuckoo didn't already use this functionality seems to
be the fact that it was created when Cuckoo didn't support version
pinning yet, i.e., it wasn't a Python package yet and as such
environments with sflock already installed without updating sflock
would've broken due to the usage of a new API endpoint.

[1]: https://github.com/pidydx/libmagicwin64