upgrade zipjail, adds max cpu/time limitations
This commit implements the long awaited requirement to limit CPU and
harddisk resources. From now on, submitted Zip bombs and alike will
affect the Cuckoo host in a limited fashion since, i.e., write a maximum
of one gigabyte (by default) to the harddisk and do this within a
maximum of two minutes (120 seconds).